주문 시 포인트 적립 삭제 주문 확정시 포인트 적립 포인트 계산 소숫점 반영 안되는 문제 수정

+++ src/main/java/com/tech/petfriends/mypage/dto/GradeDto.java
@@ -9,5 +9,5 @@ public class GradeDto {
     private int g_no;
     private String g_name;
     private int g_condition;
-    private int g_rate;
+    private float g_rate;
 }
